finish 29

+++ admin/includes/view_all_users.php
@@ -8,6 +8,10 @@
                                     <th>Email</th>
                                     <th>Role</th>
                                     <th>image</th>
+                                    <th>Admin</th>
+                                    <th>Subsriber</th>
+                                    <th>Delete</th>
+                                    <th>Edit</th>
                                 </tr>
                             </thead>
                             
@@ -52,22 +56,25 @@
         $user_id = $_GET['admin'];
         $query = "UPDATE users SET user_role='admin' WHERE user_id = $user_id";
         $result = mysqli_query($connection, $query);
-        header("Location:users.php");
         confirmQuery($result);
+        header("Location:users.php");
     }
     if(isset($_GET['subscriber'])) {
         $user_id = $_GET['subscriber'];
         $query = "UPDATE users SET user_role='subscriber' WHERE user_id = $user_id";
         $result = mysqli_query($connection, $query);
-        header("Location:users.php");
         confirmQuery($result);
+        header("Location:users.php");
     }
     if(isset($_GET['delete'])) {
-        $user_id = $_GET['delete'];
-        $query = "DELETE FROM users WHERE user_id = {$user_id}";
-        $result = mysqli_query($connection, $query);
-        header("Location:users.php");
-        confirmQuery($result);
+        if(isset($_SESSION['user_role']) && $_SESSION['user_role'] == 'admin') {
+            $user_id = mysqli_real_escape_string($connection,$_GET['delete']);
+            $query = "DELETE FROM users WHERE user_id = {$user_id}";
+            $result = mysqli_query($connection, $query);
+            confirmQuery($result);
+            header("Location:users.php");
+        }
+        
     
     }
 ?>
